deduplicate getCurrentDateString function
